requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
# sqlite3

# Telegram
python-telegram-bot>=20.0

# Web (позже)
fastapi>=0.100.0
uvicorn>=0.20.0

# Utilities
pyyaml>=6.0
loguru>=0.7.0

# Development
pytest>=7.0
black>=23.0